+++ tests/archivey/errors/test_path_traversal.py
