+++ tests/archivey/readers/test_open_compressed_stream.py
@@ -3,7 +3,6 @@ import logging
 
 import pytest
 
-from archivey.config import ArchiveyConfig
 from archivey.core import open_compressed_stream
 from archivey.exceptions import ArchiveNotSupportedError
 from archivey.types import TAR_COMPRESSED_FORMATS
@@ -12,7 +11,7 @@ from tests.archivey.sample_archives import (
     SINGLE_FILE_ARCHIVES,
     filter_archives,
 )
-from tests.archivey.testing_utils import skip_if_package_missing
+from tests.archivey.testing_utils import skip_if_package_missing, build_config
 
 BASIC_ZIP_ARCHIVE = filter_archives(BASIC_ARCHIVES, extensions=["zip"])[0]
 
@@ -28,15 +27,7 @@ logger = logging.getLogger(__name__)
 def test_open_compressed_stream_from_file(
     sample_archive, sample_archive_path, alternative_packages
 ):
-    if alternative_packages:
-        config = ArchiveyConfig(
-            use_rapidgzip=True,
-            use_indexed_bzip2=True,
-            use_python_xz=True,
-            use_zstandard=True,
-        )
-    else:
-        config = ArchiveyConfig()
+    config = build_config(alternative_packages)
 
     skip_if_package_missing(sample_archive.creation_info.format, config)
 
@@ -64,15 +55,7 @@ def test_open_compressed_stream_unsupported_format(tmp_path):
 def test_open_compressed_stream_from_stream(
     sample_archive, sample_archive_path, alternative_packages
 ):
-    if alternative_packages:
-        config = ArchiveyConfig(
-            use_rapidgzip=True,
-            use_indexed_bzip2=True,
-            use_python_xz=True,
-            use_zstandard=True,
-        )
-    else:
-        config = ArchiveyConfig()
+    config = build_config(alternative_packages)
 
     skip_if_package_missing(sample_archive.creation_info.format, config)
 
@@ -95,15 +78,7 @@ def test_open_compressed_stream_from_stream(
 def test_open_compressed_stream_from_stream_with_prefix(
     sample_archive, sample_archive_path, alternative_packages
 ):
-    if alternative_packages:
-        config = ArchiveyConfig(
-            use_rapidgzip=True,
-            use_indexed_bzip2=True,
-            use_python_xz=True,
-            use_zstandard=True,
-        )
-    else:
-        config = ArchiveyConfig()
+    config = build_config(alternative_packages)
 
     skip_if_package_missing(sample_archive.creation_info.format, config)
 
@@ -128,15 +103,7 @@ def test_open_compressed_stream_from_stream_with_prefix(
 def test_open_compressed_stream_from_archive(
     sample_archive, sample_archive_path, alternative_packages
 ):
-    if alternative_packages:
-        config = ArchiveyConfig(
-            use_rapidgzip=True,
-            use_indexed_bzip2=True,
-            use_python_xz=True,
-            use_zstandard=True,
-        )
-    else:
-        config = ArchiveyConfig()
+    config = build_config(alternative_packages)
 
     skip_if_package_missing(sample_archive.creation_info.format, config)
