+++ tests/archivey/readers/test_open_stream.py
